Dependency injection with spring and without spring
Tests have an example of DI without spring, simulating what it actually does behind the scene.
On main application there is an example of the three types of injection being made to the
controllers, property, setter and constructor, using spring @Controller annotation on controllers
and @Service on the service interface implementation, as well as the @Autowired annotation to
make the injection (on constructors its not needed).